+++ CHANGELOG.md
@@ -28,6 +28,19 @@ The general format is:
 - C from D
 
 ```
+# 2.1.1 - DATE (16/02/2022)
+
+
+### Changed
+
+- Added `volume_mounts` definition back to script template in `Task.get_script_def`.
+
+# 2.1.0 - DATE (15/02/2022)
+
+### Added
+
+- Added `TaskSecurityContext` to allow setting security settings to the task container.
+- Added `WorkflowSecurityContext` to allow setting security settings to all of the containers in the workflow.
 
 # 2.0.0 - DATE (08/02/2022)